AITRexRush/Algorithm_1/requirements.txt
numpy
selenium
pillow
pyautogui
matplotlib
python-opencv
keras
tensorflow-gpu